requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
